Make textures only set filtering if the user hasn't already set it.
This means you can do this

    var t = Textures.loadTexture("some.jpg");
    t.setParameter(gl.TEXTURE_MIN_FILTER, gl.NEAREST);

Before, the code would wait until `some.jpg` had loaded and then based
on its dimensions it would change the filtering to make it renderable
but in the example above gl.NEAREST is already fine.

Maybe it should check if the filtering is good and set it to something
that is with a warning if it's not. That way you couldn't easily
shoot yourself in the foot.